Combine final_serial and MP3_3
final_serial_2 combines the line following code from MP3_3 with the serial communication code from final_serial.

final_serial_2 and final_serial_2a cleaned up the serial print sections of the code. I created a function called serialComms() that contains all of the print information. I also created two variables (motorLeftValue and motorRightValue) that record the speed of the left and right motors respectively. these two variables are incorporated into the turnRight() and turnLeft() functions. Whenever the robot changes states, the values for the motor speeds are updated and can be pulled for serialComms().

final_serial_3 is the working version. NOTE: I did remove the line break between the left motor/sensor output and the right motor/sensor output so that I could verify that the display logic was working correctly and aligned with the robot's actual motion.